@@ -54,7 +54,7 @@
 
     <properties>
         <maven.compiler.source>10</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.target>10</maven.compiler.target>
     </properties>
 
 
@@ -66,7 +66,7 @@
                 <version>3.8.0</version>
                 <configuration>
                     <source>10</source>
-                    <target>1.8</target>
+                    <target>10</target>
                 </configuration>
                 <dependencies>
                     <dependency>
