@@ -4,8 +4,10 @@ import org.rug.data.Project;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.CopyOption;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 import java.util.Arrays;
 
 public class ArcanRunner extends ToolRunner {
@@ -16,8 +18,8 @@ public class ArcanRunner extends ToolRunner {
     /**
      * Initializes an arcan runner with the following smells CD, HL, and UD.
      */
-    public ArcanRunner(Project project, String version, String outputDir, boolean useNeo4j){
-        super("arcan");
+    public ArcanRunner(String command, Project project, String version, String outputDir, boolean useNeo4j){
+        super("arcan", command);
         var args = Arrays.asList("-p", project.getVersionedSystem().get(version).getA().toString(),
                 project.isFolderOfFoldersOfJarsProject() ? "-folderOfJars" : "-jar",
                 "-CD", "-HL", "-UD", "-CM", "-PM",
@@ -39,7 +41,7 @@ public class ArcanRunner extends ToolRunner {
     protected void postProcess(Process p) throws IOException {
         try {
             Files.move(Paths.get(getHomeDir(), "ToySystem-graph.graphml"),
-                    project.getVersionedSystem().get(version).getB());
+                    project.getVersionedSystem().get(version).getB(), StandardCopyOption.REPLACE_EXISTING);
         }catch (IOException e){
             throw new IOException("Could not move the graph file: " + e.getMessage());
         }
